Remove cosmoschangefeedcache class and merge it into cosmoschangedfeedbasedreplicationfeed
